requirements.txt
python-telegram-bot==13.15
gspread==5.12.0
requests==2.31.0
openai==1.10.0
